% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naCounter.R
\name{naCounter}
\alias{naCounter}
\title{Count all NAs in a data set}
\usage{
naCounter(data)
}
\arguments{
\item{data}{A data.frame or matrix}
}
\value{
Vector of NA counts
}
\description{
Count all NAs in a data set
}
\examples{
naCounter(airquality)

}

