% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleFromData.R
\name{sampleFromData}
\alias{sampleFromData}
\title{Sample from a data frame}
\usage{
sampleFromData(data, size, replace = TRUE, ...)
}
\arguments{
\item{data}{A data frame or matrix from which rows are
to be sampled}

\item{size}{Numeric. Number of rows to return}

\item{replace}{Logical. Sample with replacement? TRUE by default.}

\item{...}{Extra options passed to \code{\link{sample}} function.}
}
\value{
A data set of the same type as input with \code{size} rows.
}
\description{
This function has been designed to sample
from the rows of a two dimensional data set,
returning all columns of sampled rows.
}
\examples{
sampleFromData(airquality, size=10)
}
\author{
Doug Ashton
}

