% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePlot.R
\name{samplePlot}
\alias{samplePlot}
\title{plot a sample}
\usage{
samplePlot(data, size)
}
\arguments{
\item{data}{data frame}

\item{size}{numeric}
}
\value{
ggplot object
}
\description{
go away
}
\examples{
\dontrun{
samplePlot(airquality, 10)
}

}

