% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTools-package.R
\docType{data}
\name{simDat}
\alias{simDat}
\title{simDat data set}
\format{A data.frame with 3 columns
\describe{
\item{ID}{ID value}
\item{Value}{Measured value in pounds}
\item{Apples}{Logical. Do they like apples}
}}
\source{
Simulated Data
}
\usage{
simDat
}
\description{
We made some data for the package
}
\keyword{datasets}

