% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTools-package.R
\docType{package}
\name{simTools}
\alias{simTools}
\alias{simTools-package}
\title{simTools: A package for performing common simulation tasks}
\description{
This package blah blah
}
\author{
Doug Ashton \email{dashton@mango-solutions.com}
}

