% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naCounter.R
\name{sumNa}
\alias{sumNa}
\title{Count NAs in a vector}
\usage{
sumNa(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Number of NAs in x
}
\description{
Count NAs in a vector
}
\examples{
\dontrun{
  sumNa(airquality$Ozone)
}
}

